+++ inst/examples/examples-plot_jags_rhatdx.R
@@ -1,0 +1,6 @@
+
+data <- serodynamics::nepal_sees_jags_output
+
+plot_jags_Rhat(data = data,
+               iso = "HlyE_IgA",
+               strat = "typhi")